python/requirements.txt
flask
RPi.GPIO
twilio
